fix: added cjs support and fixed some compilation paths

+++ src/nodec.mjs
@@ -15,17 +15,27 @@ import { guardGoToolchain } from './guardGoToolchain.mjs';
  * @property {string} entry
  * @property {string} [name]
  * @property {string} [nodeVersion]
+ * @property {boolean} [noCleanup]
+ * @property {string} [outDir]
+ * @property {'cjs' | 'esm'} [format]
  * @property {boolean} [help]
  */
 
 function printHelp() {
   console.info(`Usage of nodec:
+--noCleanup boolean
+       If true, will leave all of the downloaded, bundled and compresses assets in a temporary
+       folder on your machine, so you can inspect the state of them
 --entry string
        (Required): the entrypoint to your JavaScript or TypeScript application
+--format string
+       which module format your JavaScript and / or TypeScript code will be compiled to. Supports 'cjs' or 'esm'. Defaults to 'esm'.
 --name string
        the final outputted filename that represents your compiled application (default "my-app")
 --node-version string
        defines the version of NodeJS that will be used when compiling your standalone executable. Must be an explicit version. Semver ranges are not supported. (default "${defaultNodeVersion}")
+--outDir
+       if set, uses this dir as the location where your binaries will be placed after compilation. Defaults to your CWD.
 --target
        one or more comma-separated os+arch compilation targets: ${Array.from(supportedOs).join(',')}
 `);
@@ -34,8 +44,15 @@ function printHelp() {
 /**
  * Cleans up the download folder after nodec executes,
  * regardless of whether it was a success or a failure
+ * @param {boolean} noCleanup
  */
-function cleanup() {
+function cleanup(noCleanup) {
+  if (noCleanup) {
+    return console.info(
+      'you opted for --no-cleanup so all artifacts have been left in the following temp folder:',
+      getTempFolder(),
+    );
+  }
   fs.removeSync(getTempFolder());
 }
 
@@ -44,8 +61,17 @@ function cleanup() {
  * @param {NodecOpts} opts
  */
 async function nodec(opts) {
+  let {
+    entry,
+    format = 'esm',
+    help,
+    name = 'my-app',
+    noCleanup = false,
+    nodeVersion = defaultNodeVersion,
+    outDir = process.cwd(),
+    target,
+  } = opts;
   try {
-    let { entry, help, name = 'my-app', nodeVersion = defaultNodeVersion, target } = opts;
     if (help) return printHelp();
 
     // first, validate the opts and / or set defaults
@@ -73,16 +99,16 @@ async function nodec(opts) {
     const downloadLocations = await Promise.all(target.map(async osname => downloadNode(nodeVersion, osname)));
 
     const bundleLocations = await Promise.all(
-      downloadLocations.map(async nodePath => bundleEntrypoint(entry, nodePath, nodeVersion)),
+      downloadLocations.map(async nodePath => bundleEntrypoint(entry, nodePath, nodeVersion, format)),
     );
 
     await Promise.all(
-      target.map(async (target, i) => compileBinary(bundleLocations[i], downloadLocations[i], target, name)),
+      target.map(async (target, i) => compileBinary(bundleLocations[i], downloadLocations[i], target, name, outDir)),
     );
 
-    cleanup();
+    cleanup(noCleanup);
   } catch (error) {
-    cleanup();
+    cleanup(noCleanup);
     console.error(error.message || error);
     process.exit(1);
   }
